@@ -24,7 +24,7 @@
     "prepare": "husky"
   },
   "engines": {
-    "node": ">= 16.0.0 < 21.0.0",
+    "node": ">= 18.0.0 < 23.0.0",
     "npm": ">= 8.x <= 10.x"
   },
   "jest": {
@@ -68,6 +68,7 @@
     "@angular/cli": "~18.2.12",
     "@angular/compiler-cli": "^18.2.13",
     "@faker-js/faker": "^9.2.0",
+    "@trivago/prettier-plugin-sort-imports": "^4.3.0",
     "@types/express": "^4.17.21",
     "@types/jasmine": "~5.1.5",
     "@types/node": "^22.10.1",
