@@ -1,8 +1,7 @@
 /* eslint-disable accessor-pairs */
-import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core'
-import { CustomerViewModel, convertToCustomerViewModel } from 'src/core/customer.viewmodel'
-import { Customer } from 'src/core/customer'
 import { NgOptimizedImage } from '@angular/common'
+import { ChangeDetectionStrategy, Component, input, output } from '@angular/core'
+import { CustomerViewModel, convertToCustomerViewModel } from 'src/core/customer.viewmodel'
 
 @Component({
   changeDetection: ChangeDetectionStrategy.OnPush,
@@ -12,17 +11,11 @@ import { NgOptimizedImage } from '@angular/common'
   templateUrl: './customer-list-item.component.html',
 })
 export class CustomerListItemComponent {
-  public customerData!: CustomerViewModel
+  public customer = input.required({ transform: convertToCustomerViewModel })
 
-  @Input({ required: true })
-  public set customer(value: Customer) {
-    this.customerData = convertToCustomerViewModel(value)
-  }
-
-  @Output()
-  public selectCustomer = new EventEmitter<CustomerViewModel>()
+  public selectCustomer = output<CustomerViewModel>()
 
   public onclick() {
-    this.selectCustomer.emit(this.customerData)
+    this.selectCustomer.emit(this.customer())
   }
 }
