@@ -1,9 +1,9 @@
+import bootstrap from './src/main.server'
 import { APP_BASE_HREF } from '@angular/common'
 import { CommonEngine } from '@angular/ssr'
 import express from 'express'
-import { fileURLToPath } from 'node:url'
 import { dirname, join, resolve } from 'node:path'
-import bootstrap from './src/main.server'
+import { fileURLToPath } from 'node:url'
 
 // The Express app is exported so that it can be used by serverless Functions.
 export function app(): express.Express {
