@@ -1,5 +1,5 @@
-import { ApplicationConfig, mergeApplicationConfig } from '@angular/core'
 import { appConfig } from './app.config'
+import { ApplicationConfig, mergeApplicationConfig } from '@angular/core'
 import { provideServerRendering } from '@angular/platform-server'
 
 const serverConfig: ApplicationConfig = {
