@@ -1,13 +1,12 @@
+import { NgClass } from '@angular/common'
 import { Component } from '@angular/core'
 import { FormsModule } from '@angular/forms'
-import { NgClass } from '@angular/common'
 import { Router } from '@angular/router'
 
 @Component({
   imports: [FormsModule, NgClass],
   selector: 'app-login',
   standalone: true,
-  styles: [],
   templateUrl: './login.component.html',
 })
 export class LoginComponent {
