@@ -1,20 +1,17 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core'
 import { ActiveDirective } from '../../directives/active.directive'
+import { ChangeDetectionStrategy, Component, signal } from '@angular/core'
 
 @Component({
   changeDetection: ChangeDetectionStrategy.OnPush,
   imports: [ActiveDirective],
   selector: 'app-navbar',
   standalone: true,
-  styles: [],
   templateUrl: './navbar.component.html',
 })
 export class NavbarComponent {
-  public isActive = false
-
-  constructor() {}
+  public isActive = signal(false)
 
   onMenuClick() {
-    this.isActive = !this.isActive
+    this.isActive.update(value => !value)
   }
 }
