@@ -1,18 +1,18 @@
-import { ChangeDetectionStrategy, Component, OnInit, signal } from '@angular/core'
-import { DetailComponent, MasterComponent, MasterDetailComponent } from 'src/app/components'
-import { AsyncDataStateType } from 'src/core/async-data-state'
 import { CustomerDetailsComponent } from '../../../components/customer-details/customer-details.component'
 import { CustomerListItemComponent } from '../../../components/customer-list-item/customer-list-item.component'
-import { CustomerType } from 'src/core/customer'
-import { CustomerViewModel } from 'src/core/customer.viewmodel'
-import { DataService } from '../services'
+import { MasterItemDirective } from '../../../directives/master-item.directive'
 import { FilterCustomerNamePipe } from '../pipes/filter-customer-name.pipe'
 import { FilterCustomerTypePipe } from '../pipes/filter-customer-type.pipe'
-import { FormsModule } from '@angular/forms'
-import { MasterCaptionComponent } from 'src/app/components/master-detail/master-caption/master-caption.component'
-import { MasterItemDirective } from '../../../directives/master-item.directive'
+import { DataService } from '../services'
 import { NgFor } from '@angular/common'
+import { ChangeDetectionStrategy, Component, OnInit, signal } from '@angular/core'
 import { toSignal } from '@angular/core/rxjs-interop'
+import { FormsModule } from '@angular/forms'
+import { DetailComponent, MasterComponent, MasterDetailComponent } from 'src/app/components'
+import { MasterCaptionComponent } from 'src/app/components/master-detail/master-caption/master-caption.component'
+import { AsyncDataStateType } from 'src/core/async-data-state'
+import { CustomerType } from 'src/core/customer'
+import { CustomerViewModel } from 'src/core/customer.viewmodel'
 
 @Component({
   changeDetection: ChangeDetectionStrategy.OnPush,
