@@ -1,8 +1,8 @@
-import { dashboardName, effects, reducers } from './store'
 import { CustomerDisplayPageComponent } from './customer-display-page/customer-display-page.component'
 import { CustomerFormsPageComponent } from './customer-forms-page/customer-forms-page.component'
 import { DashboardComponent } from './dashboard.component'
 import { DataService } from './services'
+import { dashboardName, effects, reducers } from './store'
 import { Route } from '@angular/router'
 import { provideEffects } from '@ngrx/effects'
 import { provideState } from '@ngrx/store'
