@@ -1,6 +1,6 @@
 import { AppComponent } from './app/app.component'
-import { bootstrapApplication } from '@angular/platform-browser'
 import { config } from './app/app.config.server'
+import { bootstrapApplication } from '@angular/platform-browser'
 
 const bootstrap = () => bootstrapApplication(AppComponent, config)
 
